Updated restructuring the folder and added React-route-dom
