+++ src/containers/Blog/Blog.js
@@ -1,65 +1,29 @@
 import React, { Component } from 'react';
-import axios from 'axios';
+//import axios from 'axios';
+import {Route, Link} from 'react-router-dom';
 
-import Post from '../../components/Post/Post';
-import FullPost from '../../components/FullPost/FullPost';
-import NewPost from '../../components/NewPost/NewPost';
+import NewPost from '../NewPost/NewPost';
+import Posts from '../Posts/Posts';
 import './Blog.css';
 
 class Blog extends Component {
-    
-    state = {
-        posts: [],
-        selectedPostId: null,
-        error: false
-    }
-    
-    componentDidMount () {
-        axios.get('/posts')
-            .then(response => {
-                const posts= response.data.slice(0,4);
-                const updatedPosts= posts.map(post=> {
-                    return {
-                        ...post,
-                        author : 'Diego'
-                    }
-                })
-                this.setState({posts: updatedPosts});
-            //    console.log(response);
-            })
-            .catch(error=> {
-                //console.log(error);
-                this.setState({error: true});
-            });
-    }
-
-    postSelectedHandler =(id) => {
-        this.setState({selectedPostId: id})
-    }
-
-    render () {
-        let posts= <p>Something went wrong!</p>
-        if(!this.state.error) {
-           posts= this.state.posts.map(post=> {
-                return <Post 
-                        key= {post.id} 
-                        title= {post.title} 
-                        author={post.author} 
-                        clicked={() => this.postSelectedHandler(post.id)}/>
-            });
-        }
-        
+    render() {
         return (
-            <div>
-                <section className="Posts">
-                    {posts}
-                </section>
-                <section>
-                    <FullPost id={this.state.selectedPostId}/>
-                </section>
-                <section>
-                    <NewPost />
-                </section>
+            <div className="Blog">
+                <header>
+                    <nav>
+                        <ul>
+                            <li><Link to="/">Home</Link></li>
+                            <li><Link to= {{
+                                pathname: '/new-post',
+                                hash: '#submit',
+                                search: '?quick-submit=true'
+                            }}>New Post</Link></li>
+                        </ul>
+                    </nav>
+                </header>
+                <Route path='/' exact component={Posts} />
+                <Route path='/new-post' component={NewPost} />
             </div>
         );
     }
