+++ src/containers/FullPost/FullPost.css
