+++ src/containers/FullPost/FullPost.js
@@ -22,7 +22,7 @@ class FullPost extends Component {
     deletePostHandler = ()=> {
         axios.delete(`/posts/${this.props.id}`)
         .then(response => {
-            console.log(response);
+            //console.log(response);
         })
     }
