+++ src/containers/NewPost/NewPost.js
@@ -18,7 +18,7 @@ class NewPost extends Component {
         };
         axios.post('/posts' ,data)
         .then((response) => {
-            console.log(response);
+            //console.log(response);
         });
     }
