+++ src/index.js
@@ -12,7 +12,7 @@ axios.defaults.headers.post['Content-Type'] = 'application/x-www-form-urlencoded
 // Add a request interceptor
 axios.interceptors.request.use( (request) => {
     // Do something before request is sent
-    console.log(request);
+    //console.log(request);
     return request;
   }, function (error) {
     // Do something with request error
@@ -23,7 +23,7 @@ axios.interceptors.request.use( (request) => {
 axios.interceptors.response.use(function (response) {
     // Any status code that lie within the range of 2xx cause this function to trigger
     // Do something with response data
-    console.log(response);
+    //console.log(response);
     return response;
   }, function (error) {
     // Any status codes that falls outside the range of 2xx cause this function to trigger
